Refactor: Se castea a long el id del item en FakeItemRepository.java

+++ src/main/java/com/cedaniel200/todolist/infrastructure/persistence/FakeItemRepository.java
@@ -16,7 +16,7 @@ public class FakeItemRepository implements ItemRepository {
     @Override
     public Item create(long listId, Item item) {
         items.add(item);
-        item.setId(items.size() - 1);
+        item.setId((long)items.size() - 1);
         return item;
     }
